+++ opentelemetry/src/global/internal_logging.rs
@@ -23,11 +23,19 @@ macro_rules! otel_info {
         {
             tracing::info!( name: $name, target: env!("CARGO_PKG_NAME"), "");
         }
+        #[cfg(not(feature = "internal-logs"))]
+        {
+            let _ = $name; // Compiler will optimize this out as it's unused.
+        }
     };
     (name: $name:expr, $($key:ident = $value:expr),+ $(,)?) => {
         #[cfg(feature = "internal-logs")]
         {
             tracing::info!(name: $name, target: env!("CARGO_PKG_NAME"), $($key = $value),+, "");
+        }
+        #[cfg(not(feature = "internal-logs"))]
+        {
+            let _ = ($name, $($value),+); // Compiler will optimize this out as it's unused.
         }
     };
 }
@@ -52,27 +60,23 @@ macro_rules! otel_warn {
         }
         #[cfg(not(feature = "internal-logs"))]
         {
-            #[allow(unused_variables)]
-            {
-
-            }
+            let _ = $name; // Compiler will optimize this out as it's unused.
         }
     };
     (name: $name:expr, $($key:ident = $value:expr),+ $(,)?) => {
         #[cfg(feature = "internal-logs")]
         {
             tracing::warn!(name: $name,
-                target: env!("CARGO_PKG_NAME"),
-                $($key = {
-                        $value
-                }),+,
-                ""
-        )        }
+                            target: env!("CARGO_PKG_NAME"),
+                            $($key = {
+                                    $value
+                            }),+,
+                            ""
+                    )
+        }
         #[cfg(not(feature = "internal-logs"))]
         {
-            {
-                let _ = ($name, $($value),+);
-            }
+            let _ = ($name, $($value),+); // Compiler will optimize this out as it's unused.
         }
     };
 }
@@ -95,11 +99,19 @@ macro_rules! otel_debug {
         {
             tracing::debug!(name: $name, target: env!("CARGO_PKG_NAME"),"");
         }
+        #[cfg(not(feature = "internal-logs"))]
+        {
+            let _ = $name; // Compiler will optimize this out as it's unused.
+        }
     };
     (name: $name:expr, $($key:ident = $value:expr),+ $(,)?) => {
         #[cfg(feature = "internal-logs")]
         {
             tracing::debug!(name: $name, target: env!("CARGO_PKG_NAME"), $($key = $value),+, "");
+        }
+        #[cfg(not(feature = "internal-logs"))]
+        {
+            let _ = ($name, $($value),+); // Compiler will optimize this out as it's unused.
         }
     };
 }
@@ -122,26 +134,25 @@ macro_rules! otel_error {
         {
             tracing::error!(name: $name, target: env!("CARGO_PKG_NAME"), "");
         }
-        #[allow(unused_variables)]
+        #[cfg(not(feature = "internal-logs"))]
         {
-
+            let _ = $name; // Compiler will optimize this out as it's unused.
         }
     };
     (name: $name:expr, $($key:ident = $value:expr),+ $(,)?) => {
         #[cfg(feature = "internal-logs")]
         {
             tracing::error!(name: $name,
-                target: env!("CARGO_PKG_NAME"),
-                $($key = {
-                        $value
-                }),+,
-                ""
-        )        }
+                            target: env!("CARGO_PKG_NAME"),
+                            $($key = {
+                                    $value
+                            }),+,
+                            ""
+                    )
+        }
         #[cfg(not(feature = "internal-logs"))]
         {
-            {
-                let _ = ($name, $($value),+);
-            }
+            let _ = ($name, $($value),+); // Compiler will optimize this out as it's unused.
         }
     };
 }