+++ opentelemetry/src/metrics/instruments/up_down_counter.rs
@@ -2,17 +2,12 @@ use crate::KeyValue;
 use core::fmt;
 use std::sync::Arc;
 
-use super::AsyncInstrument;
-
-/// An SDK implemented instrument that records increasing or decreasing values.
-pub trait SyncUpDownCounter<T> {
-    /// Records an increment or decrement to the counter.
-    fn add(&self, value: T, attributes: &[KeyValue]);
-}
+use super::SyncInstrument;
 
 /// An instrument that records increasing or decreasing values.
 #[derive(Clone)]
-pub struct UpDownCounter<T>(Arc<dyn SyncUpDownCounter<T> + Send + Sync>);
+#[non_exhaustive]
+pub struct UpDownCounter<T>(Arc<dyn SyncInstrument<T> + Send + Sync>);
 
 impl<T> fmt::Debug for UpDownCounter<T>
 where
@@ -28,19 +23,22 @@ where
 
 impl<T> UpDownCounter<T> {
     /// Create a new up down counter.
-    pub fn new(inner: Arc<dyn SyncUpDownCounter<T> + Send + Sync>) -> Self {
+    pub fn new(inner: Arc<dyn SyncInstrument<T> + Send + Sync>) -> Self {
         UpDownCounter(inner)
     }
 
     /// Records an increment or decrement to the counter.
     pub fn add(&self, value: T, attributes: &[KeyValue]) {
-        self.0.add(value, attributes)
+        self.0.measure(value, attributes)
     }
 }
 
 /// An async instrument that records increasing or decreasing values.
 #[derive(Clone)]
-pub struct ObservableUpDownCounter<T>(Arc<dyn AsyncInstrument<T>>);
+#[non_exhaustive]
+pub struct ObservableUpDownCounter<T> {
+    _marker: std::marker::PhantomData<T>,
+}
 
 impl<T> fmt::Debug for ObservableUpDownCounter<T>
 where
@@ -56,22 +54,10 @@ where
 
 impl<T> ObservableUpDownCounter<T> {
     /// Create a new observable up down counter.
-    pub fn new(inner: Arc<dyn AsyncInstrument<T>>) -> Self {
-        ObservableUpDownCounter(inner)
-    }
-
-    /// Records the increment or decrement to the counter.
-    ///
-    /// It is only valid to call this within a callback. If called outside of the
-    /// registered callback it should have no effect on the instrument, and an
-    /// error will be reported via the error handler.
-    pub fn observe(&self, value: T, attributes: &[KeyValue]) {
-        self.0.observe(value, attributes)
-    }
-}
-
-impl<T> AsyncInstrument<T> for ObservableUpDownCounter<T> {
-    fn observe(&self, measurement: T, attributes: &[KeyValue]) {
-        self.0.observe(measurement, attributes)
+    #[allow(clippy::new_without_default)]
+    pub fn new() -> Self {
+        ObservableUpDownCounter {
+            _marker: std::marker::PhantomData,
+        }
     }
 }