+++ opentelemetry/src/metrics/noop.rs
@@ -4,23 +4,22 @@
 //! has been set. It is expected to have minimal resource utilization and
 //! runtime impact.
 use crate::{
-    metrics::{
-        AsyncInstrument, InstrumentProvider, Meter, MeterProvider, SyncCounter, SyncGauge,
-        SyncHistogram, SyncUpDownCounter,
-    },
+    metrics::{InstrumentProvider, Meter, MeterProvider},
     KeyValue,
 };
 use std::sync::Arc;
 
+use super::instruments::SyncInstrument;
+
 /// A no-op instance of a `MetricProvider`
 #[derive(Debug, Default)]
-pub struct NoopMeterProvider {
+pub(crate) struct NoopMeterProvider {
     _private: (),
 }
 
 impl NoopMeterProvider {
     /// Create a new no-op meter provider.
-    pub fn new() -> Self {
+    pub(crate) fn new() -> Self {
         NoopMeterProvider { _private: () }
     }
 }
@@ -39,13 +38,13 @@ impl MeterProvider for NoopMeterProvider {
 
 /// A no-op instance of a `Meter`
 #[derive(Debug, Default)]
-pub struct NoopMeter {
+pub(crate) struct NoopMeter {
     _private: (),
 }
 
 impl NoopMeter {
     /// Create a new no-op meter core.
-    pub fn new() -> Self {
+    pub(crate) fn new() -> Self {
         NoopMeter { _private: () }
     }
 }
@@ -54,56 +53,19 @@ impl InstrumentProvider for NoopMeter {}
 
 /// A no-op sync instrument
 #[derive(Debug, Default)]
-pub struct NoopSyncInstrument {
+pub(crate) struct NoopSyncInstrument {
     _private: (),
 }
 
 impl NoopSyncInstrument {
     /// Create a new no-op sync instrument
-    pub fn new() -> Self {
+    pub(crate) fn new() -> Self {
         NoopSyncInstrument { _private: () }
     }
 }
 
-impl<T> SyncCounter<T> for NoopSyncInstrument {
-    fn add(&self, _value: T, _attributes: &[KeyValue]) {
-        // Ignored
-    }
-}
-
-impl<T> SyncUpDownCounter<T> for NoopSyncInstrument {
-    fn add(&self, _value: T, _attributes: &[KeyValue]) {
-        // Ignored
-    }
-}
-
-impl<T> SyncHistogram<T> for NoopSyncInstrument {
-    fn record(&self, _value: T, _attributes: &[KeyValue]) {
-        // Ignored
-    }
-}
-
-impl<T> SyncGauge<T> for NoopSyncInstrument {
-    fn record(&self, _value: T, _attributes: &[KeyValue]) {
-        // Ignored
-    }
-}
-
-/// A no-op async instrument.
-#[derive(Debug, Default)]
-pub struct NoopAsyncInstrument {
-    _private: (),
-}
-
-impl NoopAsyncInstrument {
-    /// Create a new no-op async instrument
-    pub fn new() -> Self {
-        NoopAsyncInstrument { _private: () }
-    }
-}
-
-impl<T> AsyncInstrument<T> for NoopAsyncInstrument {
-    fn observe(&self, _value: T, _attributes: &[KeyValue]) {
+impl<T> SyncInstrument<T> for NoopSyncInstrument {
+    fn measure(&self, _value: T, _attributes: &[KeyValue]) {
         // Ignored
     }
 }